Safari-609.1.11.1/LayoutTests/fast/dom/Document/document-set-title-no-child.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<body>
<script>
    description("On setting document.title to the empty string, no text node should be created");
    shouldBeEmptyString("document.title");
    evalAndLog("document.title = ''");
    shouldBeEmptyString("document.title");
    shouldBeNull("document.getElementsByTagName('title')[0].firstChild");

    document.title = "titletext";
    shouldBeEqualToString("document.title", "titletext");
    shouldBeNonNull("document.getElementsByTagName('title')[0].firstChild");

    evalAndLog("document.title = ''");
    shouldBeEmptyString("document.title");
    shouldBeNull("document.getElementsByTagName('title')[0].firstChild");

    document.title = "titletextnew";
    shouldBeEqualToString("document.title", "titletextnew");
    shouldBeNonNull("document.getElementsByTagName('title')[0].firstChild");

    document.title = "updatetitletextnew";
    shouldBeEqualToString("document.title", "updatetitletextnew");
    shouldBeNonNull("document.getElementsByTagName('title')[0].firstChild");

</script>
</body>
