LayoutTests/fast/dom/Document/document-set-title-no-child-expected.txt

On setting document.title to the empty string, no text node should be created

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.title is ""
document.title = ''
PASS document.title is ""
PASS document.getElementsByTagName('title')[0].firstChild is null
PASS document.title is "titletext"
PASS document.getElementsByTagName('title')[0].firstChild is non-null.
document.title = ''
PASS document.title is ""
PASS document.getElementsByTagName('title')[0].firstChild is null
PASS document.title is "titletextnew"
PASS document.getElementsByTagName('title')[0].firstChild is non-null.
PASS document.title is "updatetitletextnew"
PASS document.getElementsByTagName('title')[0].firstChild is non-null.
PASS successfullyParsed is true

TEST COMPLETE
